+++ prefect_firebolt/__init__.py
@@ -1,5 +1,6 @@
 from . import _version
 from .credentials import FireboltCredentials
+from .database import FireboltDatabase
 
 __version__ = _version.get_versions()["version"]
-__all__ = ["FireboltCredentials"]
+__all__ = ["FireboltCredentials", "FireboltDatabase"]